objId is size_t and compulsory
git-svn-id: file:///home/roberto/Git/svn@7149 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: da2c51a1563e4eec267fc34f4fd91f0d6eb774fe

+++ trunk/xmipp/libraries/data/normalize.cpp
@@ -613,7 +613,7 @@ void ProgNormalize::preProcess()
     }
 }
 
-void ProgNormalize::processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+void ProgNormalize::processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
 {
     Image<double> I;
     if (apply_geo)
